trading-system/services/lean-execution/main.py
#!/usr/bin/env python3

import os
import sys
import time
import logging
import redis
import json
from datetime import datetime
from decouple import config
import requests

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger(__name__)


class LeanExecutionEngine:
    def __init__(self):
        # Load configuration
        self.redis_config = {
            'host': config('REDIS_HOST', default='redis'),
            'port': config('REDIS_PORT', default='63379', cast=int),
            'password': config('REDIS_PASSWORD', default='redis_password')
        }
        
        # Broker API configuration (these would be real broker API keys in production)
        self.broker_config = {
            'api_key': config('BROKER_API_KEY', default='your-broker-api-key'),
            'api_secret': config('BROKER_API_SECRET', default='your-broker-api-secret'),
            'base_url': config('BROKER_API_URL', default='https://api.broker.com')
        }
        
        # Connect to Redis
        self.redis_client = redis.Redis(
            host=self.redis_config['host'],
            port=self.redis_config['port'],
            password=self.redis_config['password'],
            decode_responses=True
        )
        
        logger.info("Lean Execution Engine initialized")
    
    def listen_for_trades(self):
        """Listen for approved trades from Redis"""
        logger.info("Starting to listen for approved trades")
        
        pubsub = self.redis_client.pubsub()
        pubsub.subscribe('approved_trades')
        
        try:
            for message in pubsub.listen():
                if message['type'] == 'message':
                    try:
                        trade_data = json.loads(message['data'])
                        self.execute_trade(trade_data)
                    except json.JSONDecodeError as e:
                        logger.error(f"Failed to decode trade data: {e}")
                    except Exception as e:
                        logger.error(f"Error executing trade: {e}")
        
        except Exception as e:
            logger.error(f"Error in trade listener: {e}")
            raise
        finally:
            pubsub.close()
    
    def execute_trade(self, trade_signal):
        """Execute a trade based on the approved signal"""
        logger.info(f"Executing trade for {trade_signal['asset']}")
        
        try:
            # Extract trade parameters
            asset = trade_signal['asset']
            prediction = trade_signal['prediction']
            position_size = trade_signal['position_size']
            confidence = trade_signal['confidence']
            
            # Determine trade direction
            if prediction > 0:
                side = 'BUY'
            elif prediction < 0:
                side = 'SELL'
            else:
                logger.info("No trade executed - neutral prediction")
                return
            
            # Calculate order size (simplified)
            # In a real implementation, this would consider:
            # - Current portfolio value
            # - Asset price
            # - Position limits
            # - Risk management rules
            order_size = position_size * 1000  # Simplified calculation
            
            # Log trade details
            trade_details = {
                'asset': asset,
                'side': side,
                'size': order_size,
                'prediction': prediction,
                'confidence': confidence,
                'timestamp': datetime.utcnow().isoformat() + 'Z'
            }
            
            logger.info(f"Trade details: {trade_details}")
            
            # In a real implementation, this would call the broker API
            # For now, we'll just simulate the execution
            execution_result = self.simulate_trade_execution(trade_details)
            
            # Log execution result
            logger.info(f"Trade execution result: {execution_result}")
            
            # Save execution result (in a real implementation, this would go to a database)
            self.save_execution_result(execution_result)
            
        except Exception as e:
            logger.error(f"Failed to execute trade for {asset}: {e}")
            raise
    
    def simulate_trade_execution(self, trade_details):
        """Simulate trade execution (placeholder for real broker API)"""
        logger.info("Simulating trade execution")
        
        # In a real implementation, this would:
        # 1. Connect to broker API
        # 2. Place the order
        # 3. Wait for execution
        # 4. Return execution details
        
        # Simulate execution with random fill price and status
        import random
        
        execution_result = {
            'trade_id': f"TRADE_{int(time.time())}",
            'asset': trade_details['asset'],
            'side': trade_details['side'],
            'ordered_size': trade_details['size'],
            'filled_size': trade_details['size'] * random.uniform(0.9, 1.0),  # 90-100% fill rate
            'ordered_price': 100.0,  # Placeholder price
            'average_fill_price': 100.0 * random.uniform(0.99, 1.01),  # ±1% slippage
            'status': 'FILLED',
            'timestamp': datetime.utcnow().isoformat() + 'Z'
        }
        
        # Simulate execution time
        time.sleep(random.uniform(0.1, 0.5))
        
        return execution_result
    
    def save_execution_result(self, execution_result):
        """Save execution result to database (simplified)"""
        logger.info("Saving execution result")
        
        # In a real implementation, this would save to a database
        # For now, we'll just log it
        logger.info(f"Execution result saved: {execution_result}")
    
    def run(self):
        """Run the execution engine"""
        logger.info("Starting Lean execution engine")
        
        # Start listening for trades
        self.listen_for_trades()


def main():
    """Main entry point"""
    engine = LeanExecutionEngine()
    
    # Run the execution engine
    engine.run()


if __name__ == "__main__":
    main()
